backend/pyproject.toml
[project]
name = "tenpredict"
version = "0.1.0"
description = "Chain-of-effects penny stock prediction using swarm intelligence"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    # Neo4j
    "neo4j>=5.20.0",
    # LLM
    "openai>=1.40.0",
    # Data sources
    "praw>=7.7.0",
    "finnhub-python>=2.4.0",
    "feedparser>=6.0.0",
    "yfinance>=0.2.40",
    # NLP / text
    "rapidfuzz>=3.9.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Utils
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100
